impact_analysis/requirements.txt
pyserial>=3.5
numpy>=1.23
matplotlib>=3.7

